backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-cloud-aiplatform==1.36.4
google-generativeai==0.3.1
firebase-admin==6.2.0
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
httpx==0.25.1
pandas==2.1.3
numpy==1.26.2 